requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai>=1.51.0
twilio==9.2.1
requests==2.32.3
pdfminer.six==20240706
python-docx==1.1.2
pillow==10.4.0